The split powerup has been made and it has created hell
A lot of problems arise from having several balls bouncing around, most of them stem from the ball being able to go very slow because of being squished by other balls.

A way to fix this is to make sure the balls has no way to rest on the paddles.