Balls are given bounciness
The problem of minimum speed has been fixed by just making the balls bouncy with a physics material.

Though the problem of balls going through the paddle when the paddle is moving too fast is still a problem, going frame by frame, it seems to be what I expect, with the paddle going so fast, the ball ends up under the paddle in the frame its given the force to go up, causing it to hit the under side and making it go down.