Big Ball Power Up has been implemented
Power Up that makes the balls bigger has been made.

A positive consequence is how this power up stacks which is pretty cool.

A negative consequence is how if a new ball spawns while the power up is happening, the new spawned ball will not be big, but I'm keeping that because I think that causes more chaos which is cool.