Balls can now die
Balls now delete themselves when it falls down too far and on top of that, a gamesmanager has been created to announce the game over, which right now, happens when there is no more balls on screen, but should probably be changed into a life system.

Interesting observation with the split power up where the object referenced by the powerup will somehow keep its property is spawning a new ball when entering any trigger collider. This has been fixed by removing the reference when the power up is done spawning a new ball, but right now it is still unclear if that worked.