Slomo powerup has been implmented
A powerup that slows down the game has been created, has the ability to set exactly how slow the game will go and how long the effect should last.

In an attempt to make this unstackable, I have unfortunately also cause it so that when two of these are picked up, the second one just doesn't do anything. That should be fine tho.