Weird powerup bug fixed
There's been a prevalent bug that I just choose to ignore for a while and that is how a single ball would activate a powerup several times, causing it to stack excessively.

Causes of it is from how the ball's collider is made with the 'AddComponent', so since a spawned ball is made by copying the ball that activated the split powerup, they will already have a collider, but the code adds another one, causing the ball to have two colliders and when it hits a powerup, will activated it two times.

This problem just snowballs, since every newly made ball will have a +1 collider, causing the powerups to be very unstable in late game.